+++ lib/rest-core/middleware/json_request.rb
@@ -7,12 +7,10 @@ class RestCore::JsonRequest
   include RestCore::Middleware
 
   JSON_REQUEST_HEADER = {'Content-Type' => 'application/json'}.freeze
-  JSON_REQUEST_METHOD = [:post, :put, :patch]
 
   def call env, &k
     return app.call(env, &k) unless json_request(env)
-    return app.call(env, &k) unless
-      JSON_REQUEST_METHOD.include?(env[REQUEST_METHOD])
+    return app.call(env, &k) unless has_payload?(env)
 
     app.call(env.merge(
       REQUEST_HEADERS => JSON_REQUEST_HEADER.merge(env[REQUEST_HEADERS]||{}),